Fix management account config alias through ADF account management
Issue: #595

**Why?**

The management account alias cannot be configured by the account management
state machine, as the cross account access role does not have enough
permissions to create, list, or delete the aliases of the management account.

**What?**

The cross account management role that is deployed in the management account if
you adopt gets these permissions with this change.

It is assumes into the cross-account access role to list, create, and/or
delete the aliases. Hence, this change removes those policies from the
Lambda function itself. As those are not used and would only apply to the
deployment account.